requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.5.2
pydantic==2.6.4
httpx==0.24.1